add cms template and minor fixes
