--- CMS/placeholder.yaml.meta.yaml
@@ -1,6 +1,0 @@
----
-docstring: |
-    ## placeholder
-language: yaml
-tags: []
-expanded: false
