--- Components (Python)/Optional Properties/log.py.meta.yaml
@@ -1,13 +1,0 @@
----
-docstring: |
-    ## log
-    Write a custom log message from within a Python component. The message will appear in the **Logs** panel.
-    ### Properties
-    | Property | Description | Required |
-    |   ----   |     ---     |   ----   |
-    `context`|The actual message to log. Can be a string, dictionary, list, set, integer, or float. Can include Mustache syntax.|**Required**
-    `type`|The type of log. A string with no spaces.|*Optional. Default: `misc`*
-    `status`|One of `info`, `warning`, or `error`.|*Optional. Default:`info`*
-language: yaml
-tags: []
-expanded: false
