+++ Components (Python)/Utilities/meya.log.py.meta.yaml
@@ -8,24 +8,6 @@ docstring: |
     `context`|The actual message to log. Can be a string, dictionary, list, set, integer, or float. Can include Mustache syntax.|**Required**
     `type`|The type of log. A string with no spaces.|*Optional. Default: `misc`*
     `status`|One of `info`, `warning`, or `error`.|*Optional. Default:`info`*
-    ### Example
-    ```
-    from meya import Component
-
-
-    class LogComponent(Component):
-        def start(self):
-            status = "error"
-            context = {
-                "order": "chicken shawarma",
-                "sauces": [
-                        "garlic",
-                        "hot"
-                    ]
-            }
-            self.log(context, type="food", status=status)
-            return self.respond(action="next")
-    ```
 language: yaml
 tags: []
 expanded: false
