--- Templates/placeholder.yaml
@@ -1,1 +1,0 @@
-a
